+++ CHANGELOG.md
@@ -4,6 +4,11 @@ All notable changes to this package will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/)
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
+## [3.0.1] - 2021-07-22
+### Fixed
+- Perform the appropriate color space conversion for Texture Sampling sources when required.
+- Fix vertically flipped outputs on OpenGL hardware.
+
 ## [3.0.0] - 2021-06-17
 ### Changed
 - Prevent invalid GPU callback data from being written to a frame: this change skips the problematic frame and logs an error message.